Fixed a problem in converting timestamps

+++ sopstat/src/time.h
@@ -8,7 +8,7 @@
 #define GRANULARITY 3
 
 typedef struct time_stat {
-	u_long ts; /* Relative timestamp */
+	int ts; /* Relative timestamp */
 	long size[FLOWS]; /* aggragate IPlen size */
 	int pkt[FLOWS]; /* number of packets */
 	int host[FLOWS]; /* number of host */
@@ -21,6 +21,6 @@ void register_packet(time_stat *, packet_stat *, int);
 void init_time_stat(time_stat *);
 int print_time(time_stat *, char *);
 void print_time_flow(time_stat *, int);
-long timeval_difference(struct timeval a, struct timeval b);
+int timeval_difference(struct timeval a, struct timeval b);
 
 #endif /*TIME_H_*/
